CI: Update isolated job

+++ tools/ci/azure/install-posix.sh
@@ -17,7 +17,10 @@ fi
 echo "Python location: $(where python)"
 python -m pip install --upgrade pip setuptools wheel build
 python -m pip install -r requirements-dev.txt
-python -m pip uninstall numpy scipy pandas cython -y
+
+if [[ ${USE_CONDA} != "true" ]]; then
+  python -m pip uninstall numpy scipy pandas cython -y
+fi
 
 if [[ -n ${NUMPY} ]]; then CMD="$CMD==${NUMPY}"; fi;
 CMD="$CMD scipy"
